refactor(js): consolidate Jest/Vitest parsing into parse_test_output
Move Jest/Vitest JUnit XML parsing from separate module into the main
parse_test_output.py file to reduce module fragmentation. Remove
duplicate tests that were testing the now-deleted module.
